Fix SQL file migration id strings on Windows
On windows ragtime used a windows-specific path separator in a
regular expression intended to get the basename of an SQL file
migration to use as that migration's id.  However the filenames
from Java resources have platform independent paths. The defect
is the regular expression on Windows fails to produce the expected
migration id string.

In the process of making this change, non-related tests that
attempted to match multi-line output strings failed on Windows.
Multiline strings in the affected tests were changed to to use
platform-independent line endings to match output on any os.

Resolves: https://github.com/weavejester/ragtime/issues/103