Support turning off transactions
Support the optional :transactions key in a migration to control
whether to use a DB transaction in the up and/or down direction of
applying the migration.

Recognized values of the option are :up, :down, :both and true. Other
values turn off transactions in both directions. Default value is
:both.

By default, all applications of a migration (upwards or downwards) run
in a transaction.