Add migration strategy useful for continuous delivery
Blue/green deployments is a common technique used in continuous
delivery, which basically means that two versions of an application
running at the same time against the same database.

The existing raise-error migration strategy is not adequate for this
scenario as if the old version of the application is restarted, the
old version will refuse to start as it will find the new (future)
migrations applied and raise an error.

With this new "ignore-future" migration, the old version of the
application will expect that some "future" migrations have already
been applied and do not raise an error in this case.